Backend/requirements.txt
fastapi
uvicorn
prisma
asyncpg
databases[postgresql]
qrcode[pil]
pillow
google-generativeai
